+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `laravel-mailable-test` will be documented in this file
 
 - add support for Laravel 5.5
 - dropped support for older versions of Laravel
+- renamed config file from `laravel-mailable-test` to `mailable-test`
 
 ## 1.0.0 - 2017-XX-XX